add property balance test
