+++ property-authentication-deleted/index.ts
@@ -5,7 +5,7 @@ import { getPropertyByMetrics } from '../common/block-chain/utils'
 import {
 	setProcessedBlockNumber,
 	getProcessedBlockNumber,
-	getEventRecord,
+	getEventRecordThenGreaterBlockNumber,
 } from '../common/db/dao'
 import { DbConnection, Transaction } from '../common/db/common'
 import { PropertyAuthentication } from '../entities/property-authentication'
@@ -141,7 +141,7 @@ class PropertyAuthenticationDeleter extends TimerBatchBase {
 	private async getEvents(con: Connection): Promise<MetricsFactoryDestroy[]> {
 		const blockNumber = await getProcessedBlockNumber(con, this.getBatchName())
 		this.logging.infolog(`processed block number:${blockNumber}`)
-		const records = await getEventRecord(
+		const records = await getEventRecordThenGreaterBlockNumber(
 			con,
 			MetricsFactoryDestroy,
 			blockNumber + 1
